[client] refactoring/ fix component's import sections
