+++ src/components/BottomPopup.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import { editEnd } from '../modules/bookmark';
+import { editEnd } from '../actions/editBookmark';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faPlus } from '@fortawesome/free-solid-svg-icons';
 import CollectInputBox from './CollectInputBox';
