+++ src/components/CategoryReallyDeleteModal.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useDispatch } from 'react-redux';
-import { deleteCategory } from '../modules/category';
+import { deleteCategory } from '../actions/deleteCategory';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faTimes } from '@fortawesome/free-solid-svg-icons';
